+++ nextjs/src/app/reports/[id]/page.tsx
@@ -48,7 +48,7 @@ import { Button, buttonVariants } from "@/components/ui/button";
 import { useRouter } from "next/navigation";
 import spaceCase from 'to-space-case'
 import { toastPromise } from "@/lib/toast";
-import { MAP_REPORT_LAYER_ANALYTICS, ReportMap, selectedConstituencyAtom } from "@/components/report/ReportMap";
+import { MAP_REPORT_LAYER_ANALYTICS, ReportMap } from "@/components/report/ReportMap";
 import { MAP_REPORT_FRAGMENT, isConstituencyPanelOpenAtom, isDataConfigOpenAtom, isNavigatorPanelOpenAtom } from "@/lib/map";
 import { DisplayOptionsType, ReportContext, defaultDisplayOptions } from "./context";
 import { LoadingIcon } from "@/components/ui/loadingIcon";
@@ -66,6 +66,8 @@ import Navigator from "@/components/navigator/Navigator";
 type Params = {
   id: string
 }
+
+export const selectedConstituencyAtom = atom<string | null>(null)
 
 export default function Page({ params: { id } }: { params: Params }) {
   const client = useApolloClient();
